FX/requirements.txt
alpaca-py==0.15.0
asgiref==3.7.2
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
gunicorn==21.2.0
hiredis==2.3.2
pillow==10.2.0
pre-commit==3.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
sqlparse==0.4.4
